Adam: Fixed Expenses to accept decimal places - still need to work out how to convert ListArray to HashMap and back again :)

+++ app/src/main/java/com/example/metaMergeTasker/Expense.java
@@ -2,9 +2,9 @@ package com.example.metaMergeTasker;
 
 public class Expense {
     private String expenseName;
-    private int expenseAmount;
+    private double expenseAmount;
 
-    public Expense(String expenseName, int expenseAmount) {
+    public Expense(String expenseName, double expenseAmount) {
         this.expenseName = expenseName;
         this.expenseAmount = expenseAmount;
     }
@@ -17,7 +17,7 @@ public class Expense {
         this.expenseName = expenseName;
     }
 
-    public int getExpenseAmount() {
+    public double getExpenseAmount() {
         return expenseAmount;
     }
